% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create-records}
\alias{create-records}
\alias{create_records}
\title{Create}
\usage{
create_records(records, typecast = TRUE, base = NULL, table = NULL,
  quiet = !interactive())
}
\arguments{
\item{records}{A `data.frame` of records to create.}

\item{typecast}{The Airtable API will perform best-effort
automatic data conversion from string values if this is set to `TRUE`.}

\item{base}{Base, whatever this is Airtable does not document it.}

\item{table}{Name of table.}

\item{quiet}{Set to `TRUE` to print helpful messages.}
}
\value{
A list of created records.
}
\description{
Create records
}
