% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{records_to_tibble}
\alias{records_to_tibble}
\title{Convert}
\usage{
records_to_tibble(records, convert_date = TRUE)
}
\arguments{
\item{records}{Records obtained from other functions of the package.}

\item{convert_date}{Set to `TRUE` to convert dates to `POSIX`.}
}
\value{
A `tibble` containing the fields of the table together with
 the `record_id` and `record_created_time` columns.
}
\description{
Convert results to a `tibble`.
}
