% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve-records}
\alias{retrieve-records}
\alias{retrieve_record}
\alias{retrieve_records}
\title{Retrive}
\usage{
retrieve_record(record, base = NULL, table = NULL,
  quiet = !interactive())

retrieve_records(records, base = NULL, table = NULL,
  quiet = !interactive())
}
\arguments{
\item{record, records}{Record(s) to retrieve.}

\item{base}{Base, whatever this is Airtable does not document it.}

\item{table}{Name of table.}

\item{quiet}{Set to \code{TRUE} to print helpful messages.}
}
\description{
Retrive a record
}
