% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{retrieve_record}
\alias{retrieve_record}
\title{Retrive}
\usage{
retrieve_record(record, base = NULL, table = NULL,
  quiet = !interactive())
}
\arguments{
\item{record}{A record to retrieve.}

\item{base}{Base, whatever this is Airtable does not document it.}

\item{table}{Name of table.}

\item{quiet}{Set to `TRUE` to print helpful messages.}
}
\description{
Retrive a record
}
