% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{rtable}
\alias{rtable}
\title{Airtable}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtable
}
\description{
Create an rtable object.
}
\section{Methods}{

Methods sarting in \code{get_} return \code{tibble} or \code{list} objects other methods return the 
class invisibly.
\itemize{
 \item{\code{new}: create an \code{rtable}.}
 \item{\code{silent}: make the object silent, equivalent to \code{quiet=FALSE} in other functions.}
 \item{\code{set_records}: add pre-existing records to the \code{rtable} from a \code{data.frame}.}
 \item{\code{list_records}: download records from Airtable, \emph{overrides} current records.}
 \item{\code{retrieve_record}: retrieve a specific record.}
 \item{\code{retrieve_records}: retrieve specific records}
 \item{\code{create_records}: create records.}
 \item{\code{update_records}: update records.}
 \item{\code{delete_records}: delete records.}
 \item{\code{delete_record}: delete a record.}
 \item{\code{get_created}: returns created records.}
 \item{\code{get_listed}: returns records downloaded.}
 \item{\code{get_updated}: returns updated records.}
 \item{\code{get_deleted}: returns deleted records.}
 \item{\code{get_retrieved}: returns retrieved records.}
 \item{\code{refresh}: refresh records uses arguments last used \code{list_records} or defaults if never used.}
}
}

\section{Arguments}{

\itemize{
 \item{\code{new}: see \code{\link{setup}}, also takes an additional \code{list_records} argument which defaults to \code{TRUE} and collects all records in table.}
 \item{\code{silent}: takes a boolean.}
 \item{\code{set_records}: a \code{data.frame} or \code{tibble}.}
 \item{\code{list_records}: see \code{\link{list_records}}.}
 \item{\code{retrieve_record}: see \code{\link{retrieve_record}}.}
 \item{\code{retrieve_records}: see \code{\link{retrieve_records}}.}
 \item{\code{create_records}: see \code{\link{create_records}}.}
 \item{\code{update_records}: see \code{\link{update_records}}.}
 \item{\code{delete_records}: see \code{\link{delete_records}}.}
 \item{\code{delete_record}: see \code{\link{delete_record}}.}
 \item{\code{get_created}: takes \code{to_tibble} to optionally convert to a \code{tibble}.}
 \item{\code{get_listed}: takes \code{to_tibble} to optionally convert to a \code{tibble}.}
 \item{\code{get_updated}: takes \code{to_tibble} to optionally convert to a \code{tibble}.}
 \item{\code{get_deleted}: takes \code{to_tibble} to optionally convert to a \code{tibble}.}
 \item{\code{get_retrieved}: takes \code{to_tibble} to optionally convert to a \code{tibble}.}
 \item{\code{refresh}: takes \code{to_tibble} to optionally convert to a \code{tibble}.}
}
}

\keyword{datasets}
