% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\alias{get_setup}
\alias{reset_setup}
\title{Configure}
\usage{
setup(base = NULL, table = NULL, api_key = NULL)

get_setup(quiet = !interactive())

reset_setup(base = TRUE, table = TRUE, api_key = FALSE)
}
\arguments{
\item{base}{An Airtable base name.}

\item{table}{An Airtable table name.}

\item{api_key}{Your Airtable API key.}

\item{quiet}{Set to \code{TRUE} to print helpful messages.}
}
\description{
Specifies API KEY for session
}
\examples{
setup("xXXXXxxxXXXxx")
get_setup()

}
