% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_records}
\alias{update_records}
\title{Create}
\usage{
update_records(records, ids, base = NULL, table = NULL,
  quiet = !interactive())
}
\arguments{
\item{records}{A `data.frame` of records to create.}

\item{ids}{Bare column name containing record ids to update.}

\item{base}{Base, whatever this is Airtable does not document it.}

\item{table}{Name of table.}

\item{quiet}{Set to `TRUE` to print helpful messages.}
}
\value{
A list of created records.
}
\description{
Create records
}
