% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{records_to_tibble}
\alias{records_to_tibble}
\title{Convert}
\usage{
records_to_tibble(records, convert_date = TRUE, tz = Sys.timezone())
}
\arguments{
\item{records}{Records obtained from other functions of the package.}

\item{convert_date}{Set to \code{TRUE} to convert record creation time to \code{POSIXct}.}

\item{tz}{Timzone, used if \code{convert_date} is set to \code{TRUE}.}
}
\value{
A \code{tibble} containing the fields of the table together with
 the \code{record_id} and \code{record_created_time} columns.
}
\description{
Convert results to a \code{tibble}.
}
