% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{data}
\name{rtable}
\alias{rtable}
\title{Airtable}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
rtable
}
\description{
Create an rtable object.
}
\section{Methods}{

Methods sarting in `get_` return `tibble` or `list` objects other methods return the 
class invisibly.
* `new` - Create an `rtable`.
* `silent` - Make the object silent, equivalent to `quiet=FALSE` in other functions.
* `set_records` - Add pre-existing records to the `rtable` from a `data.frame`.
* `list_records` - Download records from Airtable, _overrides_ current records.
* `retrieve_record` - Retrieve a specific record.
* `retrieve_records` - Retrieve specific records.
* `create_records` - Create records.
* `update_records` - Update records.
* `delete_records` - Delete records.
* `delete_record` - Delete a record.
* `get_created` - Returns created records.
* `get_listed` - Returns records downloaded.
* `get_updated` - Returns updated records.
* `get_deleted` - Returns deleted records.
* `get_retrieved` - Returns retrieved records.
* `refresh` - Refresh records uses arguments last used `list_records` or defaults if never used.
}

\section{Arguments}{

* `new` - see \code{\link{setup}}, also takes an additional `list_records` argument which defaults to `TRUE` and collects all records in table.
* `silent` - takes `TRUE` or `FALSE`.
* `set_records` - see \code{\link{setup}}.
* `list_records` - see \code{\link{list_records}}.
* `retrieve_record` - see \code{\link{retrieve_record}}.
* `retrieve_records` - see \code{\link{retrieve_records}}.
* `create_records` - see \code{\link{create_records}}.
* `update_records` - see \code{\link{update_records}}.
* `delete_records` - see \code{\link{delete_records}}.
* `delete_record` - see \code{\link{delete_record}}.
* `get_created` - Takes `to_tibble` to optionally convert to a `tibble`.
* `get_listed` - Takes `to_tibble` to optionally convert to a `tibble`.
* `get_updated` - Takes `to_tibble` to optionally convert to a `tibble`.
* `get_deleted` - Takes `to_tibble` to optionally convert to a `tibble`.
* `get_retrieved` - Takes `to_tibble` to optionally convert to a `tibble`.
* `refresh` - No arguments.
}

\keyword{datasets}
