% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\alias{get_setup}
\title{Configure}
\usage{
setup(api_key = NULL, base = NULL, table = NULL, view = NULL)

get_setup(quiet = !interactive())
}
\arguments{
\item{api_key}{Your Airtable API key.}

\item{base}{An Airtable base name.}

\item{table}{An Airtable table name.}

\item{view}{A view name.}

\item{quiet}{Set to \code{TRUE} to print helpful messages.}
}
\description{
Specifies API KEY for session
}
\examples{
setup("xXXXXxxxXXXxx")
get_setup()

}
